requirement.txt
asgiref==3.2.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-isbn-field==0.5.2
django-multiselectfield==0.1.11
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.5
pep8==1.7.1
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.4
pycodestyle==2.5.0
pygame==1.9.6
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-stdnum==1.13
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.14.0
sqlparse==0.2.4
threadpoolctl==2.1.0
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
whitenoise==5.0.1
